+++ src/domain/type/token/payload.type.ts
@@ -1,0 +1,5 @@
+export type PayloadType = {
+  sub: number;
+  iat: number;
+  exp: number;
+};
